+++ TTKCore/musicNetworkKits/xm/musicdownloadqueryxmcommentsthread.cpp
@@ -44,6 +44,10 @@ void MusicDownLoadQueryXMCommentsThread::startSearchSong(QueryType type, const Q
 
 void MusicDownLoadQueryXMCommentsThread::startSearchSong(int offset)
 {
+    if(!m_manager)
+    {
+        return;
+    }
     deleteAll();
     m_pageTotal = 0;
